cache/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
bs4>=0.0.1
httpx>=0.23.1
pytz==2022.7
lxml>=4.9.2
# `lxml` is better. If you get error when installing lxml, you may use the below package.
# html5lib==1.1 
apscheduler==3.9.1.post1
python-dotenv